backend/requirements.txt
Django>=4.1.2
django-environ>=0.9.0
djangorestframework>=3.14.0
django-cors-headers>=3.13.0
requests>=2.27.1
sentence-transformers>=2.1.0
